[FIXED MIFOS-2410] Patch from Jakub Sławiński. https://mifos.dev.java.net/nonav/issues/showattachment.cgi/1206/issue_2410_v3.patch https://mifos.dev.java.net/issues/show_bug.cgi?id=2410 http://thread.gmane.org/gmane.comp.finance.mifos.devel/6898
git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@16198 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ application/src/main/java/org/mifos/application/customer/business/CustomerBO.java
@@ -988,7 +988,7 @@ public abstract class CustomerBO extends BusinessObject {
                 }
 
                 for (CustomerCustomFieldEntity fieldEntity : getCustomFields()) {
-                    if (fieldView.getFieldId().equals(fieldEntity.getFieldId())) {
+                    if (fieldView.getFieldId().equals(fieldEntity.getFieldId().shortValue())) {
                         fieldEntity.setFieldValue(fieldView.getFieldValue());
                     }
                 }
@@ -1151,7 +1151,7 @@ public abstract class CustomerBO extends BusinessObject {
     private void createCustomFields(final List<CustomFieldView> customFields) {
         if (customFields != null) {
             for (CustomFieldView customField : customFields) {
-                addCustomField(new CustomerCustomFieldEntity(customField.getFieldId(), customField.getFieldValue(),
+                addCustomField(new CustomerCustomFieldEntity(customField.getFieldId().intValue(), customField.getFieldValue(),
                         this));
             }
         }
